More testing

+++ src/z21Drive/actions/Z21ActionGetLocoInfo.java
@@ -24,9 +24,9 @@ public class Z21ActionGetLocoInfo extends Z21Action{
     @Override
     public void addDataToByteRepresentation(Object[] objs) {
         //Add all the data
-        byteRepresentation.add(Byte.decode("0xE3"));
-        byteRepresentation.add(Byte.decode("0xF0"));
-        byte Adr_MSB = (Byte) objs[0];
+        byteRepresentation.add((byte)(int)Integer.decode("0xE3"));
+        byteRepresentation.add((byte)(int)Integer.decode("0xF0"));
+        byte Adr_MSB = (byte)(int)(Integer)objs[0];
         byte Adr_LSB = (byte)24;
         byteRepresentation.add(Adr_MSB);
         byteRepresentation.add(Adr_LSB);
